requirements.txt
streamlit
numpy
opencv-python-headless
selenium
webdriver-manager
Pillow
numpy
pandas
cryptography
python-dotenv
streamlit-cookies-manager